+++ data_remote/src/main/java/com/z_company/entity/SectionElectric.kt
@@ -1,14 +1,15 @@
 package com.z_company.entity
 
 import com.z_company.domain.entities.route.LocoType
+import java.math.BigDecimal
 
 data class SectionElectric(
     var sectionId: String = "",
     var locoId: String = "",
     val type: LocoType = LocoType.ELECTRIC,
-    var acceptedEnergy: Double? = null,
-    var deliveryEnergy: Double? = null,
-    var acceptedRecovery: Double? = null,
-    var deliveryRecovery: Double? = null
+    var acceptedEnergy: BigDecimal? = null,
+    var deliveryEnergy: BigDecimal? = null,
+    var acceptedRecovery: BigDecimal? = null,
+    var deliveryRecovery: BigDecimal? = null
 )
